v0.2 Fixed scoreboard multiple touch issue
Issue: Rapidly tapping the -present scoreboard button- would presents it at a higher point than initially intended. 

This has been fixed using the scoreboardLauncher object that uses a screen to cover the view when the present scoreboard button gets tapped while simultaneously abstracting the responsibility, of handling the presentation and dismiss of the scoreboard, away from the splashController.  

Further improvement: Currently tapping the screen dismisses the scoreboard. A swipe gesture on the scoreboard would further complete the dismissing of the scoreboard.