v0.5 Fix: puzzleBoard issue. Added: swipeMethod
Issue: The cells in the puzzle board had unequal spacing in vertical and horizontal direction. This has been fixed by adding the "minimumInteritemSpacingForSectionAt" delegate method and setting the returned space to 0.

Cause: Even if the cell width is equal to the width of the collectionView divided by the number of cells within that collectionView, the default spacing is set to keep the cells from touching. (This means Cw != CVw / #C  if space != 0)

Added: The puzzle board collectionView correctly prints a statement with the actual swipe direction. This will be used to distinguish te different actions that need to be handled for the different swipe directions.

Next: Add an array of images to the puzzle board cells