v0.6 Building puzzleBoard test
Added Images to puzzle board tiles. Implemented logic to determine wether testImage is at either if of the four board edges. If it is, it prints a statement informing that we are at the board edge. If we are not at the board edge we print the swipe direction performed.

Note: Instead of a print statement some feedback to the user (e.g. a pulse animations on the tile, a sound or both) would be nice

Next: Rearrange the image array according to the direction in which he swipe occurs.

Note: Since this influences the UI we might need to do this on the main thread. 
Note: This probably is without animation by default so we might also need to implement some animations for this.