v0.7 Added cell move on swipe
Added:

The tiles in the puzzelboard now correctly move when swiped in a certain direction. Certain moves, e.g those that would move the empty square off the board, are forbidden.

When the elements inside the array are in a specific order the console prints a statement.

Note: Need to move all code handling the game logic concerning the arrays and its mutations, checking wether it is the solution etc. to its own object in order to maintain single responsibility.

Todo: Shuffle game array when game starts (calculate if necessary to prevent near solution game array at start)
Todo: implement audio or visual feedback when user is trying to swipe "the missing tile of the board"
Todo: Implement timer that starts/resets whenever a game starts
Todo: Implement timer stop when solution array matches game array
Todo: When timer stopped time should be stored and added to high scores