v0.8 Added Timer
Added: Foundation for timer, including pause button. Timer now starts when puzzle board is presented and resets when it is dismissed.
Added: Puzzle board launcher now has a label within the timerView that gets presented along with it. 
Todo: Optimize position for timeLabel

Changed: Scoreboard cell now has a collection of labels instead of views for holding the high scores:
Todo: Optimize position and font size for high scores

Note: Still need to connect the pause functionality of the timer to for example an in game menu/settings.
Todo: Need to find a font for the Timer that has equal decimal width to prevent the numbers from moving around