v0.9 Added image logic
Added logic that takes an image from an array of images and divides it into 16 pieces. One of the pieces then gets swapped with an "empty" image and together they fill an array. This array of images is then used to populate the puzzle tiles.

Issue: Currently there seem to be two instances of the image array. One that is visible and one that gets altered in the background when user swipes the puzzle board.

Todo: Investigate multiple instance issue and find a workaround